fix(website): isolate profile validation into import-free module [T003144]
validateProfileInput + enum consts moved from customer-crm-db.ts into the new
leaf module profile-validation.ts (no imports -> loadable under plain
tsx/node, no Vite-only import.meta.glob). customer-crm-db.ts imports and
re-exports the symbols, keeping its public API identical. BATS test now
imports the leaf module and the CONTACT_TYPES negative test got a positive
anchor (T002356-M1). Allowlist entry removed -> test back in the offline
unit gate.

Closes T003144
