+++ website/src/data/openspec-status.json
@@ -3189,6 +3189,12 @@
       "status": "plan_staged"
     }
   ],
+  "T003144": [
+    {
+      "slug": "portal-profile-bats",
+      "status": "plan_staged"
+    }
+  ],
   "T003180": [
     {
       "slug": "branch-reaper-sweep-cli",
